fix: resolve NextAuth route exports and update all import references
- Move authOptions from route file to separate lib/auth.ts file
- Update 20+ API route files to import authOptions from @/lib/auth
- Fix Next.js App Router compliance by removing config exports from route handlers
- Resolve 400 Bad Request errors on authentication-dependent endpoints

Fixes Vercel deployment failure: "authOptions is not a valid Route export field"
All API routes now properly import shared auth configuration.
