+++ src/app/api/playlists/[id]/tracks/route.ts
@@ -1,5 +1,5 @@
 import { getServerSession } from "next-auth";
-import { authOptions } from "@/app/api/auth/[...nextauth]/route";
+import { authOptions } from "@/lib/auth";
 import { ExtendedSession } from "@/types/auth";
 import { Session } from "next-auth";
 import { PrismaClient } from "@prisma/client";
