+++ src/app/api/spotify/top-tracks/route.ts
@@ -1,5 +1,5 @@
 import { getServerSession } from "next-auth";
-import { authOptions } from "../../auth/[...nextauth]/route";
+import { authOptions } from "@/lib/auth";
 import { fetchSpotifyUserTopTracks } from "@/lib/spotifyApi";
 import { ExtendedSession } from "@/types/auth";
 import { Session } from "next-auth";
