+++ src/app/api/user/route.ts
@@ -1,6 +1,6 @@
 import { getServerSession, Session } from "next-auth";
 import { PrismaClient } from "@prisma/client";
-import { authOptions } from "../auth/[...nextauth]/route";
+import { authOptions } from "@/lib/auth";
 import { ExtendedSession } from "@/types/auth";
 
 const prisma = new PrismaClient();
